God Class - Refractoring 1,2 TranscriptSequence
